Add multiple obstacles and add Java 8 compability

+++ core/src/com/mygdx/game/Obstacle.java
@@ -32,7 +32,7 @@ public class Obstacle {
         return lowerPart;
     }
 
-    public void moveHorizontally(int distance) {
+    public void moveHorizontally(float distance) {
         this.upperPart.x += distance;
         this.lowerPart.x += distance;
     }
